Added scripts to update files and run tests
The python script will create/change the necessary files.
It will read data from `data/input.json` and process the information,
creating the `.input.log` file (which will be the source of events to
lgostash) and updating the `conf/output.conf` file with the fields that
we want to test and the `scripts/run_tests.sh` file with the assertions.