Added python script to run tests
Moved away from the bash script to a python one because it's easier to
write and read.

Also, now it's possible to add several events and expectations, so that
all can be testes in 1 run.

The format of the `input.json` file needs to be the following:
```
[
  {
    "expectations": [
      {
        "[field1]": "Field 1 value"
      },
      {
        "[nested][field]": "Nested field value"
      }
    ],
    "fields": [
      "field1",
      "field2"
    ],
    "log": "log message"
  }
]
```

Notice the `[]` around non-nested fields too.